[ai] layout update
That is great, but the new button is kinda ugly. Can you make it
prettier? Maybe could you put the text, "New tab" or "Open in new tab"
in the button? I'm not sure if this is bad UI/UX practice, but could you
put the "new tab" button over the default button at the top right
corner? Any other ideas?